// src/app/features/inventario/models/requests/crear-producto-request.interface.ts
export interface SolicitudCrearProducto {
  nombre: string;
  descripcion: string;
  precio: number;
  peso: number;
  requierePreparacion: boolean;
  categoriaId: string | null;
  nuevaCategoriaNombre: string;
  buffetId: string;
  stockActual: number;
  clasificacionesSaludIds: string[];
  tiposIds: null;
  urlImagen?: string | null;
}